refactor newClient into a reusable function and rename functions
